Moved TimeUI into time_ui.py
